test/test_output/workspace_3-3/requirements.txt
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
redis==4.6.0
python-dotenv==1.0.0